gnome-post-receive-email: Handle initial push
Commit fe7b408 (Don't show commits since the beginning of time for
branch creation) introduced a check which uses the parent of the first
commit we're interested in showing.  If this hook is enabled when the
initial push is made to the repository, that commit has no parent and
the call to git rev-list yields an error (and 128 return code).  This
prevents that by checking that the parent of first_detailed_commit is a
valid ref before using it.