Handle InnoDB tables in copy-backup.py
mysqlhotcopy doesn't work for InnoDB tables. If we find any tables
that mysqlhotcopy can't handle, use mysqldump --single-transaction
instead to dump the database.

(We also use mysqldump --single-transaction for the case of mixed
tables, even though it won't give a consistent snapshot. This is
the behavior we want for bugzilla.)